requirements.txt
iexfinance
requests
ariadne==0.10.0
asgiref==3.2.3
Django==3.0.4
graphql-core==3.0.3
pytz==2019.3
sqlparse==0.3.1
starlette==0.13.2
typing-extensions==3.7.4.1
